plugins/src/main/java/com/jraska/github/client/firebase/GCloudCommands.kt
package com.jraska.github.client.firebase

object GCloudCommands {
  fun firebaseRunCommand(
    testConfiguration: TestConfiguration,
    envVars: Map<String, String>
  ): String {
    if (envVars.isEmpty()) {
      throw IllegalArgumentException("Version without env vars not supported for simplicity now")
    }

    val envVarsString = envVars.entries.map { "${it.key}=${it.value}" }.joinToString(",")

    val devicesString = testConfiguration.devices
      .joinToString(" ") { "--device " + it.firebaseCommandString() }

    return ("gcloud " +
      "firebase test android run " +
      "--app ${testConfiguration.appApkPath} " +
      "--test ${testConfiguration.testApkPath} " +
      "$devicesString " +
      "--results-dir ${testConfiguration.resultDir} " +
      "--no-performance-metrics " +
      "--num-flaky-test-attempts=1 " +
      "--timeout 3m " +
      "--environment-variables $envVarsString")
  }
}
